ENKI: change the way component attributes are stored to improve import speed ENKI: optimize some equality operations ENKI: cache size in ListProxy to avoid calls to Hibernate interceptor ENKI: remove regular expression invocation from enumeration literal parsing ENKI: remove repeated TLS look-ups
[git-p4: depot-paths = "//open/enki/": change = 10479]

+++ src/org/eigenbase/enki/jmi/impl/RefBaseObjectBase.java
@@ -142,7 +142,7 @@ public abstract class RefBaseObjectBase implements RefBaseObject
     public int hashCode()
     {
         long mofId = getMofId();
-        return (int)(mofId ^ (mofId >>> 32));
+        return ((int)mofId) ^ ((int)(mofId >>> 32));
     }
 
     /**
